app/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.5.4
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-phone-field==1.8.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
drf-yasg2==1.19.3
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.4
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.14.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.11
webencodings==0.5.1
